python/github_api/requirements.freeze.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
github3.py==3.2.0
idna==3.4
pkg_resources==0.0.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
